runtime: keep proc pin hooks in the core runtime
